清理多余基类代码
