+++ assets/scripts/verlet-assembler.ts
@@ -20,38 +20,11 @@ export default class VerletAssembler extends GTSimpleSpriteAssembler2D {
         this.indicesCount = 6 * segmentCount;
         this.floatsPerVert = 6;
         
-        //@ts-ignore
-        this._renderData = new cc.RenderData();
-        this._renderData.init(this);
-
-        // this.initLocal();
-        // this.initData();
-
-        let data = this._renderData;
-        // createFlexData支持创建指定格式的renderData
-        data.createFlexData(0, this.verticesCount, this.indicesCount, this.getVfmt());
-
-        // 顶点数固定的情况下索引不变化
-        let indices = data.iDatas[0];
-        let count = indices.length / 6;
-        for (let i = 0, idx = 0; i < count; i++) {
-            let vertextID = i * 4;
-            indices[idx++] = vertextID;
-            indices[idx++] = vertextID+1;
-            indices[idx++] = vertextID+2;
-            indices[idx++] = vertextID+1;
-            indices[idx++] = vertextID+3;
-            indices[idx++] = vertextID+2;
-        }
+        this.initData();
     }
 
     getVfmt() {
         return vfmtPosUvColorFront;
-    }
-
-    getBuffer() {
-        //@ts-ignore
-        return cc.renderer._handle.getBuffer("mesh", this.getVfmt());
     }
     
     public updateRenderData (comp: any) {
@@ -73,7 +46,7 @@ export default class VerletAssembler extends GTSimpleSpriteAssembler2D {
             let nextU = 0
 
             let floatsPerVert = this.floatsPerVert;
-            let verts = this._renderData.vDatas[0];
+            let verts = this.renderData.vDatas[0];
             let dstOffset;          // index of verts[]
             for (let i = 1; i < pointNum; i++) {
                 let lastPoint = pointList[i - 1]
@@ -81,7 +54,6 @@ export default class VerletAssembler extends GTSimpleSpriteAssembler2D {
                 nextU = lastU + gapU
 
                 // 顶点和质点一一对应
-                // same as updateVerts()
                 dstOffset = floatsPerVert * (i-1) * 4;
                 verts[dstOffset]     = posX + lastPoint.x;
                 verts[dstOffset + 1] = posY + lastPoint.y;
@@ -105,10 +77,11 @@ export default class VerletAssembler extends GTSimpleSpriteAssembler2D {
                 verts[dstOffset + 1] = posY + height + nextPoint.y;
                 verts[dstOffset + 2] = nextU;
                 verts[dstOffset + 3] = 0;
-                // dstOffset += floatsPerVert;
 
                 lastU = nextU
             }
+
+            this.updateColor(comp, null);
         }
     }
 }
